+++ experimentarium/data_react/webloader.py
@@ -4,20 +4,21 @@ import sys
 import urllib.request
 import warnings
 
-from pathlib import Path
 from subprocess import call
 from contextlib import contextmanager
 
 
-__all__ = ["DataDownloader"]
+__all__ = ["WebDataDownloader"]
 
 
 def curl(url) -> None:
+    """Terminal curl command."""
     call(f"curl --progress-bar -O {str(url)}", shell=True)
 
 
 @contextmanager
 def cd(new_dir: str) -> None:
+    """Context manager for changing directory."""
     prev_dir = os.getcwd()
     os.chdir(os.path.expanduser(new_dir))
     try:
@@ -27,42 +28,47 @@ def cd(new_dir: str) -> None:
 
 
 def is_connected() -> bool:
+    """Check if internet connection exists."""
     try:
-        urllib.request.urlopen("http://www.google.com/", timeout=1)
+        urllib.request.urlopen("http://www.google.com/", timeout=20)
         return True
     except urllib.request.URLError:
         return False
 
 
-class DataDownloader(object):
+class WebDataDownloader(object):
     def __init__(self, root: str) -> None:
+        """
+        Arguments:
+            root: Path to store downloaded data.
+                Will be created if not exists.
+        """
         self.root = os.path.abspath(root)
-        if not os.path.exists(self.root):
-            os.mkdir(self.root)
-        with open(
-            os.path.join(Path(__file__).resolve().parents[0], "dataset2url.json")
-        ) as f:
-            self.dataset2url = json.load(f)
+        with open(os.path.join(os.path.dirname(__file__), "dir2url.json")) as f:
+            self.dir2url = json.load(f)
 
     def download(self, name: str) -> bool:
-        if name not in self.dataset2url:
-            raise ValueError(f"Unknown dataset {name}.")
+        if name not in self.dir2url:
+            raise ValueError(f"Unknown benchmark {name}.")
         if not is_connected():
             return False
         if not os.path.exists(self.root):
             os.mkdir(self.root)
-        dataset = self.dataset2url[name]
-        dataset_root = os.path.join(self.root, name)
-        if not os.path.exists(dataset_root):
-            os.mkdir(dataset_root)
-        with cd(dataset_root):
-            for url in dataset["urls"]:
+
+        benchmark = self.dir2url[name]
+        benchmark_root = os.path.join(self.root, name)
+        if not os.path.exists(benchmark_root):
+            os.mkdir(benchmark_root)
+
+        with cd(benchmark_root):
+            for url in benchmark["urls"]:
                 _, filename = os.path.split(url)
-                file_path = os.path.join(dataset_root, filename)
+                file_path = os.path.join(benchmark_root, filename)
+
                 if os.path.exists(file_path):
                     warnings.warn(f"File alredy exists, no download.\n")
                 else:
-                    sys.stdout.write(f"Downloading file to {dataset_root}.\n")
+                    sys.stdout.write(f"Downloading file to {benchmark_root}.\n")
                     sys.stdout.flush()
                     curl(url)
         return True
