--- experiments/demo/params.json
@@ -1,16 +1,0 @@
-{
-    "model": "sla",
-    "benchmarks": [
-        "banknotes",
-        "pendigits_4_9",
-        "fashion_mnist_4_9"
-    ],
-    "log": "True",
-    "n_states": 1,
-    "lsizes": [
-        10
-    ],
-    "config_path": "demo/config.py",
-    "data_root": "demo/.data",
-    "log_root": "demo/"
-}
